RP2350 and SDK 2.0.0 changes

+++ adc/adc_console/adc_console.c
@@ -17,7 +17,7 @@ void printhelp() {
 void __not_in_flash_func(adc_capture)(uint16_t *buf, size_t count) {
     adc_fifo_setup(true, false, 0, false, false);
     adc_run(true);
-    for (int i = 0; i < count; i = i + 1)
+    for (size_t i = 0; i < count; i = i + 1)
         buf[i] = adc_fifo_get_blocking();
     adc_run(false);
     adc_fifo_drain();
